[edit] completed section 06

+++ 06-variables/05-type-conversion/exercises/05-convert-and-fix-5/main.go
@@ -29,5 +29,5 @@ func main() {
 	max := int16(1000)
 
 	// FIX THE CODE HERE
-	fmt.Println(int8(max) + min)
+	fmt.Println(int32(max) + int32(min))
 }
